% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adni2.R
\docType{data}
\name{adni2}
\alias{adni2}
\title{adni2}
\format{
A data frame with 317 rows and 7 columns:
\describe{
\item{RID}{Participant ID}
\item{DX.bl}{The disease class label}
\item{FDG}{Numeric, value of FDG}
\item{AV45}{Numeric, value of AV45}
\item{ABETA}{Numeric, value of ABETA}
\item{TAU.x}{Numeric, value of TAU from CSF}
\item{PTAU}{Numeric, value of PTAU from CSF}
}
}
\source{
This is a subset of ADNI2 dataset, available at https://adni.loni.usc.edu
}
\description{
Description of adni2.
}
