% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/True_values.R
\name{AUCofunc}
\alias{AUCofunc}
\title{R function that calculates the true values of AUCo when distribution is known}
\usage{
AUCofunc(k1, k2, distribution, arg1, arg2)
}
\arguments{
\item{k1}{number of subclasses in main class-1}

\item{k2}{number of subclasses in main class-2}

\item{distribution}{the distribution of marker value follows Normal or Gamma}

\item{arg1}{if distribution is normal input mean parameters of all subclasses in a vector, if gamma input shape parameters}

\item{arg2}{if distribution is gamma input variance parameter, if gamma input rate parameters}
}
\value{
The true value of AUCo under given distribution and parameters
}
\description{
R function that calculates the true values of AUCo when distribution is known
}
