% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orderstatistics.R
\name{cdf_min_given_max_partial_upper}
\alias{cdf_min_given_max_partial_upper}
\title{R function that calculates the conditional probability of minimum greater than y_min given maximum equals to y_max of random variables (upper tail probability of minimum given maximum)}
\usage{
cdf_min_given_max_partial_upper(y_min, y_max, distribution, arg1, arg2)
}
\arguments{
\item{y_min}{the value of y_min}

\item{y_max}{the value of y_max}

\item{distribution}{the distribution of marker value follows Normal or Gamma}

\item{arg1}{if distribution is normal input mean parameters of all subclasses in a vector, if gamma input shape parameters}

\item{arg2}{if distribution is normal input variance parameter, if gamma input rate parameters}
}
\value{
The conditional probability of minimum given maximum of random variables
}
\description{
R function that calculates the conditional probability of minimum greater than y_min given maximum equals to y_max of random variables (upper tail probability of minimum given maximum)
}
