% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orderstatistics.R
\name{cdf_min_max_partial}
\alias{cdf_min_max_partial}
\title{R function that calculates the partial of joint probability of min and max over max of NIND random variables}
\usage{
cdf_min_max_partial(y_min, y_max, distribution, arg1, arg2)
}
\arguments{
\item{y_min}{the value of y_min}

\item{y_max}{the value of y_max}

\item{distribution}{the distribution of marker value follows Normal or Gamma}

\item{arg1}{if distribution is normal input mean parameters of all subclasses in a vector, if gamma input shape parameters}

\item{arg2}{if distribution is normal input variance parameter, if gamma input rate parameters}
}
\value{
The partial of joint probablity of min and max over max
}
\description{
R function that calculates the partial of joint probability of min and max over max of NIND random variables
}
