% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orderstatistics.R
\name{cdf_order_r}
\alias{cdf_order_r}
\title{R function that calculates the probability of r-th order statistics of normal random variables (CDF of r-th order statistics)}
\usage{
cdf_order_r(x, r, distribution, arg1, arg2)
}
\arguments{
\item{x}{the value of x}

\item{r}{r-th order statistics}

\item{distribution}{the distribution of marker value follows Normal or Gamma}

\item{arg1}{if distribution is normal input mean parameters of all subclasses in a vector, if gamma input shape parameters}

\item{arg2}{if distribution is normal input variance parameter, if gamma input rate parameters}
}
\value{
The probability of r-th order statistics of random variables smaller or equal to x
}
\description{
R function that calculates the probability of r-th order statistics of normal random variables (CDF of r-th order statistics)
}
