% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Empirical_est.R
\name{CI.func}
\alias{CI.func}
\title{R function that calculates percentile confidence interval given an array of estimates}
\usage{
CI.func(x)
}
\arguments{
\item{x}{an array of calculated estimates}
}
\value{
The percentile confidence interval of given values
}
\description{
This function provides percentile confidence interval
}
