% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Order_statistics.R
\name{F_min_given_max_partial_gamma_upper}
\alias{F_min_given_max_partial_gamma_upper}
\title{R function that calculates the conditional probability of minimum greater than y_min given maximum equals to y_max of gamma random variables (upper tail of conditional probability of minimum given maximum)}
\usage{
F_min_given_max_partial_gamma_upper(y_min, y_max, shape, rate)
}
\arguments{
\item{y_min}{the value of y_min}

\item{y_max}{the value of y_max}

\item{shape}{the vector of shape parameters of gamma random variables}

\item{rate}{the vector of rate parameters of gamma random variables}
}
\value{
The conditional probability of minimum given maximum of gamma random variables
}
\description{
R function that calculates the conditional probability of minimum greater than y_min given maximum equals to y_max of gamma random variables (upper tail of conditional probability of minimum given maximum)
}
