% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Order_statistics.R
\name{F_min_given_max_partial_normal_upper}
\alias{F_min_given_max_partial_normal_upper}
\title{R function that calculates the conditional probability of minimum greater than y_min given maximum equals to y_max of normal random variables (upper tail probability of minimum given maximum)}
\usage{
F_min_given_max_partial_normal_upper(y_min, y_max, mu, sd)
}
\arguments{
\item{y_min}{the value of y_min}

\item{y_max}{the value of y_max}

\item{mu}{the vector of mean parameters of normal random variables}

\item{sd}{the vector of variance parameters of normal random variables}
}
\value{
The conditional probability of minimum given maximum of normal random variables
}
\description{
R function that calculates the conditional probability of minimum greater than y_min given maximum equals to y_max of normal random variables (upper tail probability of minimum given maximum)
}
