% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Order_statistics.R
\name{F_min_max_partial_gamma}
\alias{F_min_max_partial_gamma}
\title{R function that calculates the partial of joint probability of min and max over max of NIND gamma random variables}
\usage{
F_min_max_partial_gamma(y_min, y_max, shape, rate)
}
\arguments{
\item{y_min}{the value of y_min}

\item{y_max}{the value of y_max}

\item{shape}{the vector of shape parameters of gamma random variables}

\item{rate}{the vector of rate parameters of gamma random variables}
}
\value{
The partial of joint probablity of min and max over max
}
\description{
R function that calculates the partial of joint probability of min and max over max of NIND gamma random variables
}
