% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Order_statistics.R
\name{F_min_max_partial_normal}
\alias{F_min_max_partial_normal}
\title{R function that calculates the partial of joint probability of min and max over max of NIND normal random variables}
\usage{
F_min_max_partial_normal(y_min, y_max, mu, sd)
}
\arguments{
\item{y_min}{the value of y_min}

\item{y_max}{the value of y_max}

\item{mu}{the vector of mean parameters of normal random variables}

\item{sd}{the vector of variance parameters of normal random variables}
}
\value{
The partial of joint probablity of min and max over max
}
\description{
R function that calculates the partial of joint probability of min and max over max of NIND normal random variables
}
