% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Orderstatistics.R
\name{f_order_max}
\alias{f_order_max}
\title{R function that calculates the probability density of maximum of NIND random variables (PDF)}
\usage{
f_order_max(y_max, distribution, arg1, arg2)
}
\arguments{
\item{y_max}{the value of y_max}

\item{distribution}{the distribution of marker value follows Normal or Gamma}

\item{arg1}{if distribution is normal input mean parameters of all subclasses in a vector, if gamma input shape parameters}

\item{arg2}{if distribution is normal input variance parameter, if gamma input rate parameters}
}
\value{
The probability density of maximum of random variables
}
\description{
R function that calculates the probability density of maximum of NIND random variables (PDF)
}
