% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Order_statistics.R
\name{f_order_max_gamma}
\alias{f_order_max_gamma}
\title{R function that calculates the probability density of maximum of gamma random variables (PDF)}
\usage{
f_order_max_gamma(y_max, shape, rate)
}
\arguments{
\item{y_max}{the value of y_max}

\item{shape}{the vector of shape parameters of gamma random variables}

\item{rate}{the vector of rate parameters of gamma random variables}
}
\value{
The probability density of maximum of gamma random variables
}
\description{
R function that calculates the probability density of maximum of gamma random variables (PDF)
}
