% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Order_statistics.R
\name{f_order_max_normal}
\alias{f_order_max_normal}
\title{R function that calculates the probability density of maximum of NIND normal random variables (PDF)}
\usage{
f_order_max_normal(y_max, mu, sd)
}
\arguments{
\item{y_max}{the value of y_max}

\item{mu}{the vector of mean parameters of normal random variables}

\item{sd}{the vector of variance parameters of normal random variables}
}
\value{
The probability density of maximum of normal random variables
}
\description{
R function that calculates the probability density of maximum of NIND normal random variables (PDF)
}
