% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Order_statistics.R
\name{F_order_r_gamma}
\alias{F_order_r_gamma}
\title{R function that calculates the probability of r-th order statistics of gamma random variables (CDF of r-th order statistics)}
\usage{
F_order_r_gamma(x, shape, rate, r)
}
\arguments{
\item{x}{the value of x}

\item{shape}{the vector of shape parameters of gamma random variables}

\item{rate}{the vector of rate parameters of gamma random variables}

\item{r}{r-th order statistics}
}
\value{
The probability of r-th order statistics of gamma random variables smaller or equal to x
}
\description{
R function that calculates the probability of r-th order statistics of gamma random variables (CDF of r-th order statistics)
}
