% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Order_statistics.R
\name{F_order_r_normal}
\alias{F_order_r_normal}
\title{R function that calculates the probability of r-th order statistics of normal random variables (CDF of r-th order statistics)}
\usage{
F_order_r_normal(x, mu, sd, r)
}
\arguments{
\item{x}{the value of x}

\item{mu}{the vector of mean parameters of normal random variables}

\item{sd}{the vector of variance parameters of normal random variables}

\item{r}{r-th order statistics}
}
\value{
The probability of r-th order statistics of normal random variables smaller or equal to x
}
\description{
R function that calculates the probability of r-th order statistics of normal random variables (CDF of r-th order statistics)
}
