% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Empirical_est.R
\name{get_max_min_permutations}
\alias{get_max_min_permutations}
\title{R function for obtaining all combinations of maximum and minimum of a given dataset}
\usage{
get_max_min_permutations(df)
}
\arguments{
\item{df}{Given dataset, in list}
}
\value{
A list of all combinations of maximum and minimum of df
}
\description{
R function for obtaining all combinations of maximum and minimum of a given dataset
}
