% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/True_values.R
\name{HUM_standard}
\alias{HUM_standard}
\title{R function to calculate the standardized HUMcm under given structure}
\usage{
HUM_standard(value, num_sub)
}
\arguments{
\item{value}{the value of HUMcm}

\item{num_sub}{the vector of number of subclasses in each main class}
}
\value{
The standardized HUMcm
}
\description{
R function to calculate the standardized HUMcm under given structure
}
