% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/True_values.R
\name{humc_min}
\alias{humc_min}
\title{R function that calculates the minimum of HUMcm under given structure}
\usage{
humc_min(num_sub)
}
\arguments{
\item{num_sub}{the vector of number of subclasses in each main class}
}
\value{
the minimum of HUMcm
}
\description{
R function that calculates the minimum of HUMcm under given structure
}
