% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{rocc_curve}
\alias{rocc_curve}
\title{R function for plotting the overall ROC curve and chance curve}
\usage{
rocc_curve(k1, k2, distribution, arg1, arg2)
}
\arguments{
\item{k1}{number of subclasses in main class-1}

\item{k2}{number of subclasses in main class-2}

\item{distribution}{the distribution of marker value follows Normal or Gamma}

\item{arg1}{if distribution is normal input mean parameters of all subclasses in a vector, if gamma input shape parameters}

\item{arg2}{if distribution is gamma input variance parameter, if gamma input rate parameters}
}
\value{
The overall ROC curve and chance curve
}
\description{
R function for plotting the overall ROC curve and chance curve
}
