% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots2.R
\name{rocc_curve_emp}
\alias{rocc_curve_emp}
\title{R function for plotting the empirical compound ROC curve and chance curve}
\usage{
rocc_curve_emp(dat, num_sub)
}
\arguments{
\item{dat}{values in list, each element represents biomarker values for a disease group, ordered in ascending severity}

\item{num_sub}{a vector of number of subclasses in each subclass}
}
\value{
The empirical compound ROC curve and chance curve
}
\description{
R function for plotting the empirical compound ROC curve and chance curve
}
