% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plots.R
\name{rocc_surface}
\alias{rocc_surface}
\title{R function for plotting the compound ROC surface and chance surface}
\usage{
rocc_surface(k1, k2, k3, distribution, arg1, arg2)
}
\arguments{
\item{k1}{number of subclasses in main class-1}

\item{k2}{number of subclasses in main class-2}

\item{k3}{number of subclasses in main class-3}

\item{distribution}{the distribution of marker value follows Normal or Gamma}

\item{arg1}{if distribution is normal input mean parameters of all subclasses in a vector, if gamma input shape parameters}

\item{arg2}{if distribution is gamma input variance parameter, if gamma input rate parameters}
}
\value{
The compound ROC surface and chance surface
}
\description{
R function for plotting the compound ROC surface and chance surface
}
