% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/True_values.R
\name{HUMC_fourclass}
\alias{HUMC_fourclass}
\title{R function that calculates the true values of HUMcm when distribution is known}
\usage{
HUMC_fourclass(distribution, arg1, arg2, num_sub)
}
\arguments{
\item{distribution}{the distribution of marker value follows Normal or Gamma}

\item{arg1}{if distribution is normal input mean parameters of all subclasses in a vector, if gamma input shape parameters}

\item{arg2}{if distribution is gamma input variance parameter, if gamma input rate parameters}

\item{num_sub}{the vector of number of subclasses in each main class}
}
\value{
The true value of HUMcm under given distribution and parameters
}
\description{
R function that calculates the true values of HUMcm when distribution is known
}
